added new functionality to ui
