--- Billing/tests/conftest.py
@@ -1,6 +1,0 @@
-
-import os
-import tempfile
-
-import pytest
-
